vault/backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0
requests>=2.32,<3.0
gunicorn>=22.0,<23.0
django-q2>=1.6
anthropic>=0.40
drf-spectacular>=0.27
